Move PersonalizeDevfileConfig to Interactive backend
