+++ pkg/init/backend/interface.go
@@ -22,4 +22,7 @@ type InitBackend interface {
 
 	// PersonalizeName updates a devfile name, depending on the flags
 	PersonalizeName(devfile parser.DevfileObj, flags map[string]string) error
+
+	// PersonalizeDevfileConfig updates the env vars, and URL endpoints
+	PersonalizeDevfileConfig(devfile parser.DevfileObj) error
 }
